requirements.txt
ultralytics>=8.3.0
numpy
opencv-python
PyYAML
plotly>=5.18.0
